requirements.txt
PyYAML>=3.0
tqdm>=4.0
pystalk<0.3
